% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnpj.R
\name{add_digits}
\alias{add_digits}
\title{Add leading zeros in cnpjs}
\usage{
add_digits(.cnpj)
}
\arguments{
\item{.cnpj}{cnpjs}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function adds leading zeros in cnpjs
}
\examples{
add_digits("213565398")
}
