% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnpj-package.R
\docType{package}
\name{cnpj-package}
\alias{cnpj}
\alias{cnpj-package}
\title{cnpj: Tools for handling cnpj data}
\description{
This package provides tools for handling cnpj data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/arturhgq/cnpj}
  \item \url{https://arturhgq.github.io/cnpj/}
  \item Report bugs at \url{https://github.com/arturhgq/cnpj/issues}
}

}
\author{
\strong{Maintainer}: Artur Quirino \email{contact@arturhgq.com} (\href{https://orcid.org/0000-0001-8071-6348}{ORCID})

}
\keyword{internal}
